- Bug fix: Legacy transformer options start at page 1, where as image magick and the local transforms start at 0;

+++ src/test/java/org/alfresco/repo/thumbnail/ThumbnailServiceImplTest.java
@@ -233,8 +233,8 @@ public class ThumbnailServiceImplTest extends BaseAlfrescoSpringTest
     {
         ImageTransformationOptions options = new ImageTransformationOptions();
         PagedSourceOptions pagedSourceOptions = new PagedSourceOptions();
-        pagedSourceOptions.setStartPageNumber(new Integer(1));
-        pagedSourceOptions.setEndPageNumber(new Integer(1));
+        pagedSourceOptions.setStartPageNumber(new Integer(2));
+        pagedSourceOptions.setEndPageNumber(new Integer(2));
         options.addSourceOptions(pagedSourceOptions);
         
         ThumbnailDefinition thumbnailDefinition = new ThumbnailDefinition(MimetypeMap.MIMETYPE_PDF, options, "doclib_2");
